requirements.txt
prettytable==0.7.2
python-dateutil==2.7.2
six==1.11.0
